# api/openapi.yaml
openapi: 3.1.0
info:
  title: Chat Platform API
  version: "1.0.0"
  description: >-
    Multi-region, real-time chat infrastructure. This spec covers the
    customer-facing REST API only (organizations, apps, credentials,
    end-users, channels, messages, reactions, polls, read state, blocks) —
    it does not cover the /dashboard/* endpoints, which exist solely to
    power the first-party console and are not part of the public contract.


    Two callers use this API, and every operation is tagged `x-sdk-scope`
    to say which one it's for:


    - **server** — your backend, authenticating with an App's key/secret
      (HTTP Basic) or an org-admin bearer token. Creates apps, credentials,
      and end-users. Should never run in a browser or mobile client — these
      credentials are long-lived secrets.
    - **client** — an end-user's own device (web/mobile), authenticating
      with the short-lived bearer token your backend receives from
      `POST /users` and hands to that end-user. Sends messages, joins
      channels, reacts, reads.


    Real-time delivery (typing, live message push) is a separate WebSocket
    service (`/connect` on the gateway) and is not describable in OpenAPI —
    see each SDK's `realtime` module, which is hand-written rather than
    generated.
  license:
    name: MIT
servers:
  - url: https://api.example.com
    description: Production (replace with your deployed API's base URL)
  - url: http://localhost:8081
...
security: []
paths:
...
  /channels:
    post:
      operationId: createChannel
      tags: [channels]
      x-sdk-scope: client
      summary: Create a channel
      description: The channel's home region is wherever it's created — the caller becomes its first member.
      security: [{ userBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 128
      responses:
        "201":
          description: Channel created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    orgBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The org-admin token from `POST /organizations`, or a dashboard
        session token — both resolve to the same org identity server-side.
        Server-side use only.
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The end-user token from `POST /users`' response. Safe to hand to
        an end-user's own client — it can only act as that one end-user.
    appBasicAuth:
      type: http
      scheme: basic
      description: >-
        An app's API key as the username, its secret as the password (see
        `POST /apps/{app_id}/credentials`). Server-side use only — never
        ship these to a browser or mobile client. The one place this is
        still sent directly is `POST /apps/token`; every other server-scoped
        call uses the short-lived Bearer token that endpoint mints instead
        — see `appBearer`.
    appBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The token from `POST /apps/token` (itself authenticated with
        `appBasicAuth`). Carries this App's identity for `POST /users` so
        your backend sends its raw secret once per token lifetime (1h)
        instead of on every end-user it creates. Re-exchange whenever it
        expires — server-side use only, same as `appBasicAuth`.
...
  responses:
    BadRequest:
      description: The request body or a parameter failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    QuotaExceeded:
      description: A plan limit (not a rate limit) was reached — e.g. max apps, max channels, max channel members.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Channel:
      type: object
      properties:
        channel_id: { type: string, format: uuid }
        name: { type: string }
        home_region: { type: string, enum: [eu, us, asia] }
        virtual_shard: { type: integer }
      required: [channel_id, name, home_region, virtual_shard]
...